pyproject.toml
[project]
name = "pod2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.3.0",
    "click>=8.1.8",
    "librosa>=0.10.2.post1",
    "pyannote-audio>=3.3.2",
    "pydantic>=2.10.6",
    "ruff>=0.9.6",
    "soundfile>=0.13.1",
    "torch>=2.6.0",
    "transformers>=4.48.3",
]


[tool.uv]
dev-dependencies = [
    "datasets>=3.3.0",
    "huggingface-hub[cli]>=0.28.1",
]

[tool.uv.workspace]
members = ["podstar"]
